consolidate GameState
