+++ src/state.js
@@ -1,6 +1,10 @@
 export const GameState = {
   money: 10.00,
   love: 10,
+  queue: [],
+  activeCustomer: null,
+  wanderers: [],
+  sparrows: [],
   spawnTimer: null,
   falconActive: false,
   gameOver: false,
